% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadElevationsMap.R
\name{downloadElevationsMap}
\alias{downloadElevationsMap}
\title{Check if map is in local}
\usage{
downloadElevationsMap(
  MapaBase,
  pais = "",
  region = "",
  provincia = "",
  comarca = "",
  localidad = ""
)
}
\arguments{
\item{MapaBase}{Map}

\item{crs_mapabase}{Contains the CRS of the map to their parameters filtering}
}
\value{
MapaBase Rivers Map
}
\description{
Check if map is in local
}
