% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadMap.R
\name{downloadMap}
\alias{downloadMap}
\title{Download Map if it not exist in local}
\usage{
downloadMap(pais = "", nivel_solicitado)
}
\arguments{
\item{pais}{country selected}

\item{nivel_solicitado}{level of detail. 1,2,3 or 4}
}
\value{
downloadMap Return the map
}
\description{
Download Map if it not exist in local
}
