% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadRiverMap.R
\name{downloadRiverMap}
\alias{downloadRiverMap}
\title{Check if map is in local}
\usage{
downloadRiverMap(MapaBase)
}
\arguments{
\item{MapaBase}{Map}
}
\value{
MapaBase Rivers Map
}
\description{
Check if map is in local
}
