% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lockLimitDownloadCountry.R
\name{lockLimitDownloadCountry}
\alias{lockLimitDownloadCountry}
\title{Check if map is in local}
\usage{
lockLimitDownloadCountry(pais = "")
}
\arguments{
\item{pais}{country selected}

\item{nivel}{level of detail. 1,2,3 or 4}
}
\value{
MapaBase is the file with the MAP if this exists
}
\description{
Check if map is in local
}
