% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lockSelection.R
\name{lockSelection}
\alias{lockSelection}
\title{Check if map is in local}
\usage{
lockSelection(nivel)
}
\arguments{
\item{nivel}{level of detail. 1,2,3 or 4}

\item{pais}{country selected}
}
\value{
MapaBase is the file with the MAP if this exists
}
\description{
Check if map is in local
}
