% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_functions.R
\name{observeResetButton}
\alias{observeResetButton}
\title{This function reset all values and restart all maps}
\usage{
observeResetButton(session, input)
}
\arguments{
\item{session}{Contents of sesion}

\item{input}{Contents of input}
}
\value{
#No returns
}
\description{
This function reset all values and restart all maps
}
