% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderTittleMap.R
\name{renderTittleMap}
\alias{renderTittleMap}
\title{Check if map is in local}
\usage{
renderTittleMap(
  input,
  output,
  seleccion_pais,
  seleccion_region,
  seleccion_provincia,
  seleccion_comarca,
  seleccion_localidad
)
}
\arguments{
\item{input}{country selected}

\item{output}{level of detail. 1,2,3 or 4}

\item{seleccion_pais}{level of detail. 1,2,3 or 4}

\item{seleccion_region}{level of detail. 1,2,3 or}

\item{seleccion_provincia}{level of detail. 1,2,3 or 4}

\item{seleccion_comarca}{level of detail. 1,2,3 or 4}

\item{seleccion_localidad}{level of detail. 1,2,3 or 4}
}
\value{
#No returns,,. It prints in screen result of TITTLE
}
\description{
Check if map is in local
}
